requirements.txt
actionlib==1.12.1
angles==1.9.12
attrs==24.2.0
blinker==1.9.0
bondpy==1.8.5
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
catkin==0.7.29
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
controller_manager==0.18.4
controller_manager_msgs==0.18.4
cv_bridge==1.13.1
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
dynamic_reconfigure==1.6.5
exceptiongroup==1.2.2
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
greenlet==3.1.1
h11==0.14.0
idna==3.10
image_geometry==1.13.1
interactive-markers==1.11.5
itsdangerous==2.2.0
Jinja2==3.1.4
joint_state_publisher==1.12.15
kdl-parser-py==1.13.3
laser_geometry==1.6.7
MarkupSafe==3.0.2
message_filters==1.14.13
outcome==1.3.0.post0
packaging==24.2
PySocks==1.7.1
python-dotenv==1.0.1
python_qt_binding==0.4.4
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.32.3
resource_retriever==1.12.7
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-image-view==0.4.17
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rviz==1.13.30
selenium==4.26.1
sensor-msgs==1.12.8
smach==2.0.3
smach_ros==2.0.3
smclib==1.8.5
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
topic_tools==1.14.13
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urdfdom-py==0.4.6
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xacro==1.13.20
